Fix dock
git-svn-id: file:///.data/svn_convert/svn@662 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/libicq/icqclient.cpp
@@ -428,7 +428,7 @@ void ICQClientPrivate::idle()
     factory->idle();
     time_t now;
     time(&now);
-    if (now >= check_time + CHECK_INTERVAL){
+    if ((unsigned)now >= check_time + CHECK_INTERVAL){
         checkBirthDay();
         check_time = now;
     }
